Add priority_queue implementation
This commit adds the implementation of a priority queue data structure using a binary tree of nodes. The priority_queue class includes functions for adding elements, removing the element with the highest priority, retrieving the element with the highest priority, checking if the queue is empty, and getting the size of the queue. The implementation also includes a print function to display the contents of the priority queue.

The priority_queue class is defined in the priority_queue.h header file, and the implementation is in the priority_queue.cpp source file. The main.cpp file contains a sample program that allows users to interact with the priority queue by adding elements, removing the element with the highest priority, displaying the element with the highest priority, checking if the queue is empty, and getting the size of the queue.

Author: [Author Name]
Date: [Commit Date]